Long Short-Term Memory Training/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-storage==1.25.0
numpy
keras
pandas
tensorflow==2.14.0
scikit-learn>=0.24.2